Hackathon 2/requirements.txt
Flask>=2.0
python-dotenv
mysql-connector-python
requests
openai

